Lose State (WIP)
The game can now be lost as enemies damage the base. No lose screen yet.

+++ Assets/Scripts/UI/HUD/BaseHealthDisplay.cs
@@ -16,7 +16,7 @@ public class BaseHealthDisplay : MonoBehaviour
 
     void Update()
     {
-        healthBar.value = baseHealth.currentHealth / baseHealth.maxHealth;
+        healthBar.value = (float)baseHealth.currentHealth / baseHealth.maxHealth;
 
         healthText.text = "Health: " + baseHealth.currentHealth + "/" + baseHealth.maxHealth;
     }
